Change the directories of the specific page components
