+++ src/app/(home)/_components/CoreFeatures.tsx
