+++ src/app/(home)/_components/NewProvidersSection.tsx
