+++ src/app/(main)/categories/_components/CategoryCard.tsx
