+++ src/app/(main)/providers/[username]/_components/QuickInfo.tsx
