+++ src/app/(main)/providers/[username]/_components/SingleProviderBanner.tsx
