+++ src/app/(main)/service/[id]/_components/Info.tsx
