+++ src/app/(main)/service/[id]/_components/LocationAndDurationInfo.tsx
