+++ src/app/(main)/service/[id]/_components/ProductQuickInfo.tsx
