+++ src/app/(main)/service/[id]/_components/ServiceProductBanner.tsx
