--- src/components/pages/admin/Sidebar/Desktop.tsx
@@ -1,43 +1,0 @@
-import LogoutButton from "@/components/buttons/LogoutButton";
-import sidebarLinks from "@/utils/constants/adminSidebarLinks";
-import Image from "next/image";
-import Link from "next/link";
-import { SidebarProps } from "./types";
-
-export default function DesktopSidebar({ user }: SidebarProps) {
-  const { name, email, avatarImageURL } = user;
-  return (
-    <aside className="hidden lg:flex flex-col bg-primary-100 overflow-hidden p-8">
-      <div className="flex items-center gap-4">
-        <Image
-          src={avatarImageURL}
-          width={40}
-          height={40}
-          className="rounded-full bg-white"
-          alt=""
-        />
-        <div className="text-sm">
-          <p className="font-bold text-lg">{name}</p>
-          <p>{email}</p>
-        </div>
-      </div>
-      <nav className="flex-1">
-        <ul className="flex flex-col mt-12 gap-8 text-sm">
-          {sidebarLinks.map(({ href, title }) => (
-            <li key={href}>
-              <Link
-                className="bg-white shadow-sm rounded-md p-3 block hover:bg-white/70"
-                href={href}
-              >
-                {title}
-              </Link>
-            </li>
-          ))}
-        </ul>
-      </nav>
-      <div>
-        <LogoutButton type="admin" />
-      </div>
-    </aside>
-  );
-}
