--- src/components/pages/admin/Sidebar/types.ts
@@ -1,7 +1,0 @@
-export type SidebarProps = {
-  user: {
-    email: string;
-    name: string;
-    avatarImageURL: string;
-  };
-};
